refactor(hub): self-contained enclave — contract types hoisted, importCek behind barrel
kernel/enclave/** no longer type-imports from with-commit/crdt,
with-party/team, or with-audit/sealed-record: CrdtMode/CrdtState/
CrdtStrategy, RecipientSealer, and SealedCekBinding/SealedCekDeliveryEnvelope
move to kernel/types.ts (contract types), with the original modules
re-exporting them so no external importer breaks. The one raw
subtle.importKey outside the enclave (sealed-record's CEK import) now goes
through a new importCek barrel export. scripts/check-architecture.mjs's
enclave-barrel-only check gains the reverse assertion: kernel/enclave/**
importing a with-* path now fails by name.
